refactor: renamed ScrollButton file to pascal case
